+++ mocks/electronStoreMock.js
@@ -1,0 +1,3 @@
+// Allows jest to test files that import the electron-store package.
+
+module.exports = class MockStore {};