+++ src/Device/DeviceSelector/DeviceSelector.jsx
@@ -38,9 +38,8 @@ import React, { useCallback, useEffect, useState } from 'react';
 import {
     bool, exact, func, object,
 } from 'prop-types';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 
-import { devices as devicesSelector } from '../deviceReducer';
 import {
     deselectDevice,
     selectDevice,
@@ -65,8 +64,6 @@ const DeviceSelector = ({
     onDeviceIsReady = noop,
 }) => {
     const dispatch = useDispatch();
-    const devices = useSelector(devicesSelector);
-
     const [deviceListVisible, setDeviceListVisible] = useState(false);
 
     const doDeselectDevice = useCallback(
@@ -107,10 +104,10 @@ const DeviceSelector = ({
         <div className={`core19-device-selector ${deviceListVisible ? 'device-list-visible' : ''}`}>
             <SelectorButton
                 deviceListVisible={deviceListVisible}
-                setDeviceListVisible={setDeviceListVisible}
+                toggleDeviceListVisible={() => setDeviceListVisible(!deviceListVisible)}
                 doDeselectDevice={doDeselectDevice}
             />
-            { deviceListVisible && <DeviceList devices={devices} doSelectDevice={doSelectDevice} />}
+            { deviceListVisible && <DeviceList doSelectDevice={doSelectDevice} />}
 
             <DeviceSetup />
         </div>