+++ src/persistentStore.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2015 - 2017, Nordic Semiconductor ASA
+/* Copyright (c) 2015 - 2020, Nordic Semiconductor ASA
  *
  * All rights reserved.
  *
@@ -34,46 +34,12 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React from 'react';
-import { bool, node } from 'prop-types';
-import { displayedDeviceName } from '../deviceInfo/deviceInfo';
-import deviceShape from './deviceShape';
-import DeviceIcon from './DeviceIcon';
-import { FavoriteIndicator } from './Favorite';
+import Store from 'electron-store';
 
-import './basic-device-info.scss';
+export const store = new Store({ name: 'pc-nrfconnect-shared' });
 
-const DeviceName = ({ device }) => (
-    <div className="name">{displayedDeviceName(device)}</div>
-);
-DeviceName.propTypes = {
-    device: deviceShape.isRequired,
-};
+export const persistNickname = (serialNumber, nickname) => (store.set(`${serialNumber}.name`, nickname));
+export const getPersistedNickname = serialNumber => (store.get(`${serialNumber}.name`, ''));
 
-const DeviceSerialNumber = ({ device }) => (
-    <div className="serial-number">{device.serialNumber}</div>
-);
-DeviceSerialNumber.propTypes = {
-    device: deviceShape.isRequired,
-};
-
-const BasicDeviceInfo = ({ device, whiteBackground, additionalToggle }) => (
-    <div className="basic-device-info">
-        <DeviceIcon device={device} whiteBackground={whiteBackground} />
-        <div className="details">
-            <DeviceName device={device} />
-            <DeviceSerialNumber device={device} />
-        </div>
-        <div className="toggles">
-            <FavoriteIndicator device={device} />
-            {additionalToggle}
-        </div>
-    </div>
-);
-BasicDeviceInfo.propTypes = {
-    device: deviceShape.isRequired,
-    whiteBackground: bool.isRequired,
-    additionalToggle: node.isRequired,
-};
-
-export default BasicDeviceInfo;
+export const persistIsFavorite = (serialNumber, value) => (store.set(`${serialNumber}.fav`, value));
+export const getPersistedIsFavorite = serialNumber => (store.get(`${serialNumber}.fav`, false));